requirements.txt
streamlit
torch
diffusers
transformers
safetensors
Pillow
requests
